fortify fpr
